Bugfix #503 next button disables early
the math that's deciding if the next button should be enabled was missing a few things that were causing the button to disable early in the situation where more than 1 slide shows at once. 
- Adding a case for `left` to the cellAlign switch statement. If the base `slidesToShow` here is 1.25, we can assume based off the existing code that we need to offset the lastSlide by 1. 
- the `center` case needs to just look at the last slide offset, not subtract one. All we need is the slidesToShow's nearest integer. 
-  the `right` case we need to find the nearest integer that's higher than the number at hand and then subtract 2 from it to account for the space to the left. 

All this makes more sense (to me) in context with when it'd actually be present - so if more than 1 slide is getting shown we round the currentSlide and slidesShowing to the nearest integer (this is where that switch statement comes into play) added with the currentSlide.  
- if `seenSlides` is equal or greater than the `totalSlides` available then the button is disabled.